resourcemonitor: more tests
add tests to the resourcemonitor package, using mock
for PodResourcesListerClient.
This brings the coverage to >= 80%, even though it covers
only the bare bones of the expected usage scenarios.

Minor fixes along the way to make the code easier to test
without changing the functionalities.

Bump the deps to topologyinfo v0.0.1 to pull a fix
needed for testability.

Signed-off-by: Francesco Romani <fromani@redhat.com>